Generate and run all migrations to create game schema
The rails app was generated without a database, therefore a
database was created. The database did not contain any of the tables
necessary for how the app/game is to be setup. The game requires the
use of 7 tables total, 3 of which are join tables, so a separate
migration was generated for each table and then migrated immediately
after checking each for the correct input and output.