Fix user association with game
The games/users correlation was not working when tested with seeds
in the console and kept demanding a join table be used. However,
one was not necessary, but you can only use the association of
'has and belongs to many' for join table associations. Changed the
users association to 'has many through group', which works now to
call game.users