Fix game association with user
The game association for user was broken and kept demanding a join
table to be made for the association, but that wasn't necessary and
we did not want to add another table. The association 'has and belongs
to many' can only be used for join tables so the game association for
users was changed to just a regular 'has many'.